Add files via upload
This is a project that was worked on both personally and as part of a group from an association called Data analytics elites or DAElites for short. This project was the analysis conducted on a bank marketing campaign dataset of which data preprocessing and analysis were conducted and at the end of the project a marketing dashboard was created using python (streamlit library) to be able to monitor the performance of the campaign of different job roles.